+++ barretenberg/cpp/src/barretenberg/ultra_honk/relation_correctness.test.cpp
@@ -1,5 +1,4 @@
 #include "barretenberg/flavor/goblin_translator.hpp"
-#include "barretenberg/honk/composer/ultra_composer.hpp"
 #include "barretenberg/honk/proof_system/permutation_library.hpp"
 #include "barretenberg/proof_system/library/grand_product_library.hpp"
 #include "barretenberg/relations/auxiliary_relation.hpp"
@@ -10,6 +9,7 @@
 #include "barretenberg/relations/permutation_relation.hpp"
 #include "barretenberg/relations/relation_parameters.hpp"
 #include "barretenberg/relations/ultra_arithmetic_relation.hpp"
+#include "barretenberg/ultra_honk/ultra_composer.hpp"
 #include <gtest/gtest.h>
 
 using namespace proof_system::honk;
