+++ barretenberg/cpp/src/barretenberg/ultra_honk/ultra_composer.hpp
@@ -4,11 +4,11 @@
 #include "barretenberg/honk/proof_system/goblin_merge/merge_verifier.hpp"
 #include "barretenberg/honk/proof_system/protogalaxy_prover.hpp"
 #include "barretenberg/honk/proof_system/protogalaxy_verifier.hpp"
-#include "barretenberg/honk/proof_system/ultra_prover.hpp"
-#include "barretenberg/honk/proof_system/ultra_verifier.hpp"
 #include "barretenberg/proof_system/composer/composer_lib.hpp"
 #include "barretenberg/srs/global_crs.hpp"
 #include "barretenberg/sumcheck/instance/prover_instance.hpp"
+#include "barretenberg/ultra_honk/ultra_prover.hpp"
+#include "barretenberg/ultra_honk/ultra_verifier.hpp"
 
 namespace proof_system::honk {
 template <UltraFlavor Flavor> class UltraComposer_ {
