+++ barretenberg/cpp/src/barretenberg/ultra_honk/ultra_composer.test.cpp
@@ -1,7 +1,6 @@
-#include "barretenberg/honk/composer/ultra_composer.hpp"
+#include "barretenberg/ultra_honk/ultra_composer.hpp"
 #include "barretenberg/common/serialize.hpp"
 #include "barretenberg/ecc/curves/bn254/fr.hpp"
-#include "barretenberg/honk/proof_system/ultra_prover.hpp"
 #include "barretenberg/numeric/uint256/uint256.hpp"
 #include "barretenberg/proof_system/circuit_builder/ultra_circuit_builder.hpp"
 #include "barretenberg/proof_system/library/grand_product_delta.hpp"
@@ -10,6 +9,7 @@
 #include "barretenberg/relations/permutation_relation.hpp"
 #include "barretenberg/relations/relation_parameters.hpp"
 #include "barretenberg/sumcheck/sumcheck_round.hpp"
+#include "barretenberg/ultra_honk/ultra_prover.hpp"
 #include <cstddef>
 #include <cstdint>
 #include <gtest/gtest.h>
